@@ -1,21 +1,21 @@
 name: spine_client
 description: Dart-based library for client applications of Spine-based systems.
-version: 1.7.2
+version: 1.7.3
 homepage: https://spine.io
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.13.0 <3.0.0'
 
 dependencies:
-  protobuf: ^1.0.0
-  http: ^0.12.0+2
-  uuid: ^2.0.2
-  firebase: ^5.0.4
-  fixnum: ^0.10.9
-  sprintf: ^4.0.2
-  optional: ^3.1.0
+  protobuf: ^2.0.0
+  http: ^0.13.3
+  uuid: ^3.0.4
+  firebase: ^9.0.1
+  fixnum: ^1.0.0
+  sprintf: ^6.0.0
+  optional: ^6.0.0-nullsafety.2
 
 dev_dependencies:
-  pedantic: ^1.7.0
-  test: ^1.6.0
-  dartdoc: ^0.30.2
+  pedantic: ^1.11.0
+  test: ^1.17.4
+  dartdoc: ^0.43.0
