@@ -1,22 +1,22 @@
 name: dart_code_gen
 description: A command-line tool which generates Dart code for Protobuf type registries.
-version: 1.7.2
+version: 1.7.3
 homepage: https://spine.io
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.13.0 <3.0.0'
 
 dependencies:
-  args: 1.5.2
-  protobuf: ^1.0.0
-  fixnum: ^0.10.9
-  code_builder: ^3.2.0
-  dart_style: 1.3.1
-  optional: ^3.1.0
+  args: 2.1.0
+  protobuf: ^2.0.0
+  fixnum: ^1.0.0
+  code_builder: ^4.0.0
+  dart_style: ^2.0.1
+  optional: ^6.0.0-nullsafety.2
 
 dev_dependencies:
-  pedantic: ^1.8.0
-  test: ^1.6.0
+  pedantic: ^1.11.0
+  test: ^1.17.4
 
 executables:
   dart_code_gen: main
