@@ -3,12 +3,12 @@ description: Tests for package `spine_client`.
 publish_to: none
 
 environment:
-  sdk: '>=2.5.0 <3.0.0'
+  sdk: '>=2.13.0 <3.0.0'
 
 dependencies:
   spine_client:
     path: ../../client
 
 dev_dependencies:
-  pedantic: ^1.7.0
-  test: ^1.6.0
+  pedantic: ^1.11.0
+  test: ^1.17.4
