@@ -50,7 +50,10 @@ class Subscription<T extends GeneratedMessage> {
     Subscription._(this.subscription, Stream<T> itemAdded)
         : _itemAdded = _checkBroadcast(itemAdded),
           _closed = false {
-        subscription.catchError((e) => unsubscribe());
+        subscription.catchError((e) {
+            unsubscribe();
+            return pb.Subscription.getDefault();
+        });
     }
 
     /// Shows if this subscription is still active or already closed.
@@ -141,7 +144,7 @@ class EventSubscription<T extends GeneratedMessage> extends Subscription<Event> 
 
     /// A stream of typed event messages.
     Stream<T> get eventMessages => events
-        .map((event) => unpack(event.message));
+        .map((event) => unpack(event.message) as T);
 }
 
 Future<String> _nodePath(Future<FirebaseSubscription> subscription) =>
@@ -150,7 +153,6 @@ Future<String> _nodePath(Future<FirebaseSubscription> subscription) =>
 
 Stream<String> _nodePathStream(Future<String> futureValue) =>
     futureValue.asStream().asBroadcastStream();
-
 
 Stream<T> _checkBroadcast<T>(Stream<T> stream) {
     if (!stream.isBroadcast) {
